+++ test/unit/with-xstate-service.spec.mjs
@@ -6,13 +6,15 @@ import sinon from 'sinon'
 test.beforeEach(t => {
   t.context = createWrapper(
     withXStateService({
+      status: 0,
       _xstateService: {
         status: 0,
-        subscribe: sinon.stub().returns(sinon.spy()),
-        start: sinon.spy(),
-        stop: sinon.spy(),
-        send: sinon.spy()
-      }
+        state: 'test'
+      },
+      state: 'test',
+      subscribe: sinon.stub().returns(sinon.spy()),
+      start: sinon.spy(),
+      send: sinon.spy()
     }, {
       mounted: sinon.spy(),
       updated: sinon.spy(),
@@ -22,36 +24,36 @@ test.beforeEach(t => {
 })
 
 test('set-up subscription and start service', t => {
-  t.truthy(t.context.fsmService._xstateService.subscribe.called)
-  t.truthy(t.context.fsmService._xstateService.start.called)
+  t.truthy(t.context.fsm.subscribe.called)
+  t.truthy(t.context.fsm.start.called)
 })
 
 test('send action', t => {
-  t.context.send('LOADING')
-  t.truthy(t.context.fsmService._xstateService.send.called)
-  t.is(t.context.fsmService._xstateService.send.getCall(0).args[0], 'LOADING')
+  t.context.fsm.send('LOADING')
+  t.truthy(t.context.fsm.send.called)
+  t.is(t.context.fsm.send.getCall(0).args[0], 'LOADING')
 })
 
 test('stop and start service with mounted', t => {
-  t.context.fsmService._xstateService.status = 1
+  t.context.fsm.status = 1
   t.context.removed()
-  t.context.fsmService._xstateService.status = 0
+  t.context.fsm.status = 0
   t.context.mounted()
-  t.truthy(t.context.fsmService._xstateService.start.called)
+  t.truthy(t.context.fsm.start.called)
 })
 
 test('stop and start service with updated', t => {
-  t.context.fsmService._xstateService.status = 1
+  t.context.fsm.status = 1
   t.context.removed()
-  t.context.fsmService._xstateService.status = 0
+  t.context.fsm.status = 0
   t.context.updated()
-  t.truthy(t.context.fsmService._xstateService.start.called)
+  t.truthy(t.context.fsm.start.called)
 })
 
 test('start service with updated if already running', t => {
-  t.context.fsmService._xstateService.status = 0
+  t.context.fsm.status = 0
   t.context.updated()
-  t.truthy(t.context.fsmService._xstateService.start.called)
+  t.truthy(t.context.fsm.start.called)
 })
 
 test('state changes', t => {
@@ -60,12 +62,16 @@ test('state changes', t => {
     withXStateService({
       _xstateService: {
         status: 1,
-        subscribe (callback) {
-          subscribeCallback = callback
-        },
-        start () {},
-        stop () {}
-      }
+        state: 'test',
+      },
+      status: 1,
+      state: 'test',
+      subscribe (callback) {
+        subscribeCallback = callback
+        return () => {}
+      },
+      start () {},
+      send () {}
     }, {})
   )
   wrapper._processRender = sinon.spy()